% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datrand}
\alias{datrand}
\title{Simulated data set}
\format{
A data frame with 24999 rows and 4 variables:
\describe{
\item{\code{id}}{integer. A patient identifier}
\item{\code{Y}}{integer. A binary outcome}
\item{\code{time}}{double. A time-varying covariate}
\item{\code{binary}}{double. A time-invariant covariate}
}
}
\usage{
datrand
}
\description{
A simulated data set. Data were created using fixed marginal mean parameters (beta0, beta1, beta2, beta3) = (-1.85, -0.15, 1.00, 0.15) and association parameters (gamma, sigma) = (1.5, 0.0). These data were created assuming an autocorrelation dependence structure.
}
\keyword{datasets}
