% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.GH.R
\name{get.GH}
\alias{get.GH}
\title{Compute Gauss-Hermite quadrature rule}
\usage{
get.GH(q, scale_abscissa = sqrt(2), scale_weight = 1/sqrt(pi))
}
\arguments{
\item{q}{Order of the Gauss-Hermite quadrature rule to compute}

\item{scale_abscissa}{Fixed number}

\item{scale_weight}{Fixed number}
}
\value{
A list with the following components:
\describe{
\item{z}{Nodes}
\item{w}{Quadrature Weights}
}
}
\description{
Compute Gauss-Hermite quadrature rule
}
\keyword{internal}
