% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madras}
\alias{madras}
\title{Madras Longitudinal Schizophrenia Study: Thought Disorder Subset}
\format{
A data frame with 922 rows and 5 variables:
\describe{
\item{\code{id}}{integer. An indicator for thought disorders}
\item{\code{thought}}{integer. COLUMN_DESCRIPTION}
\item{\code{month}}{integer. Months since hospitalization}
\item{\code{gender}}{integer. An indicator for female gender}
\item{\code{age}}{double. An indicator for age-at-onset \eqn{>= 20} years}
}
}
\source{
Peter Diggle, Patrick J. Heagerty, Kung-Yee Liang, and Scott L. Zeger. Analysis of longitudinal data. Oxford University Press, 2002.
}
\usage{
madras
}
\description{
\code{madras} contains a subset of the data from the Madras Longitudinal Schizophrenia Study, which collected monthly symptom data on 86 schizophrenia patients after their initial hospitalization. The primary question of interest is whether subjects with an older age-at-onset tend to recover more or less quickly, and whether female patients recover more or less quickly. Recovery is measured by a reduction in the presentation of symptoms.
}
\keyword{datasets}
